manager/requirements.txt
Flask
requests
pyyaml
pydantic>=1.8.2
